file-repository/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
sqlalchemy
aiofiles
Pillow